+++ src/serial.rs
@@ -576,8 +576,13 @@ where
     /// the triggered events via [`Serial::triggered_events`].
     ///
     /// Returns `None` if the hardware is busy.
+    ///
+    /// ## Embedded HAL
+    ///
+    /// To have a more managed way to read from the serial use the [`embeded_hal::serial::Read`]
+    /// trait implementation.
     #[doc(alias = "RDR")]
-    pub fn raw_read(&self) -> Option<u8> {
+    pub fn read_data_register(&self) -> Option<u8> {
         if self.usart.isr.read().busy().bit_is_set() {
             return None;
         }
@@ -667,6 +672,43 @@ where
         for event in events.iter() {
             self.configure_interrupt(event, true);
         }
+    }
+
+    /// Check if an interrupt is configured for the [`Event`]
+    #[inline]
+    pub fn is_interrupt_configured(&self, event: Event) -> bool {
+        match event {
+            Event::TransmitDataRegisterEmtpy => self.usart.cr1.read().txeie().is_enabled(),
+            Event::CtsInterrupt => self.usart.cr3.read().ctsie().is_enabled(),
+            Event::TransmissionComplete => self.usart.cr1.read().tcie().is_enabled(),
+            Event::ReceiveDataRegisterNotEmpty => self.usart.cr1.read().rxneie().is_enabled(),
+            Event::ParityError => self.usart.cr1.read().peie().is_enabled(),
+            Event::LinBreak => self.usart.cr2.read().lbdie().is_enabled(),
+            Event::NoiseError | Event::OverrunError | Event::FramingError => {
+                self.usart.cr3.read().eie().is_enabled()
+            }
+            Event::Idle => self.usart.cr1.read().idleie().is_enabled(),
+            Event::CharacterMatch => self.usart.cr1.read().cmie().is_enabled(),
+            Event::ReceiverTimeout => self.usart.cr1.read().rtoie().is_enabled(),
+            // Event::EndOfBlock => self.usart.cr1.read().eobie().is_enabled(),
+            // Event::WakeupFromStopMode => self.usart.cr3.read().wufie().is_enabled(),
+        }
+    }
+
+    /// Check which interrupts are enabled for all [`Event`]s
+    #[cfg(feature = "enumset")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "enumset")))]
+    #[inline]
+    pub fn configured_interrupts(&mut self) -> EnumSet<Event> {
+        let mut events = EnumSet::new();
+
+        for event in EnumSet::<Event>::all().iter() {
+            if self.is_interrupt_configured(event) {
+                events |= event;
+            }
+        }
+
+        events
     }
 
     /// Check if an interrupt event happend.
@@ -907,7 +949,7 @@ where
     /// up to the interrupt handler.
     ///
     /// To read out the content of the read register without internal error handling, use
-    /// [`Serial::raw_read()`].
+    /// [`Serial::read()`].
     /// ...
     // -> According to this API it should be skipped.
     fn read(&mut self) -> nb::Result<u8, Error> {